+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 * Added Scene Composer to generate basic GUI's with a visual editor.
 * Added submenu controls in dialogs to clean up interface.
+* Added duplicate option for Tinted and Tiled Drawables.
 * Cleaned up Drawables Dialog. Popups for Drawable settings and labels describing what type of Drawable it is.
 * Drawables can be hidden in the Drawables dialog to prevent them from being shown with the default filter. They can be unhidden by deactivating the filter and deselcting "hidden".
 * Fixed crash when adding a custom placeholder drawable to a project that is set to keep resources relative.